test(#5): update tests for DOM-based stage-crew behavior; extract CanvasDrop helper; add jsdom dev dep
